feat(confluence): add confluence_download_pages MCP tool and reuse downloader in Teammate
- Extracted recursive Confluence page/attachment download logic into
  ConfluencePageDownloader so it can be shared.
- Added MCP tool confluence_download_pages(urlStrings, outputPath, depth,
  downloadAttachments) for standalone pre-download.
- Refactored CliExecutionHelper.writeConfluencePagesFile to delegate to
  ConfluencePageDownloader, ensuring Teammate and the MCP tool behave identically.
- Tested end-to-end on AI TEST and Template Story pages: 12 pages downloaded
  recursively including attachments.
