+++ dmtools-core/src/main/java/com/github/istin/dmtools/teammate/CliExecutionHelper.java
@@ -12,14 +12,9 @@ import com.github.istin.dmtools.common.utils.CommandLineUtils;
 import com.github.istin.dmtools.common.utils.IOUtils;
 import com.github.istin.dmtools.common.utils.PropertyReader;
 import com.github.istin.dmtools.atlassian.confluence.Confluence;
-import com.github.istin.dmtools.atlassian.confluence.ConfluenceStorageMarkdown;
-import com.github.istin.dmtools.atlassian.confluence.model.Content;
-import io.github.furstenheim.CopyDown;
+import com.github.istin.dmtools.atlassian.confluence.ConfluencePageDownloader;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
 
 import java.io.File;
 import java.io.IOException;
@@ -28,11 +23,8 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Queue;
 import java.util.Set;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
@@ -200,10 +192,8 @@ public class CliExecutionHelper {
      * {@code confluenceDepth} levels. Attachments are downloaded for every fetched page unless
      * {@code confluenceAttachments} is {@code false}.
      * <p>
-     * Uses {@link Confluence#parseUris} to detect URLs, {@link Confluence#contentByUrl} /
-     * {@link Confluence#findContent} to resolve pages, and
-     * {@link Confluence#downloadPageAttachments} for attachments — reusing the same patterns as
-     * {@code ConfluenceMermaidIndexIntegration}.
+     * Delegates to {@link ConfluencePageDownloader} so that Teammate and the
+     * {@code confluence_download_pages} MCP tool use the exact same download logic.
      *
      * @param textContent           Any text that may contain Confluence URLs (e.g. full ticket text)
      * @param inputFolderPath       The base input folder (e.g. {@code input/PROJ-123})
@@ -224,259 +214,11 @@ public class CliExecutionHelper {
                     seedUrls.size(), confluenceDepth, confluenceAttachments);
 
             Path confluenceFolder = inputFolderPath.resolve("confluence");
-            Files.createDirectories(confluenceFolder);
-
-            Queue<PageLevel> queue = new LinkedList<>();
-            Set<String> processedIds = new HashSet<>();
-            Set<String> processedUrls = new HashSet<>();
-
-            // Seed queue with pages linked directly from the ticket text (level 0)
-            for (String url : seedUrls) {
-                try {
-                    Content page = confluence.contentByUrl(url);
-                    processedUrls.add(url);
-                    if (page != null && page.getId() != null) {
-                        queue.add(new PageLevel(page, 0));
-                    } else {
-                        logger.warn("Confluence returned null Content for {}", url);
-                    }
-                } catch (Exception e) {
-                    logger.warn("Could not fetch Confluence seed page {} (skipping): {}", url, e.getMessage());
-                }
-            }
-
-            int written = 0;
-            int index = 0;
-            while (!queue.isEmpty()) {
-                PageLevel current = queue.poll();
-                Content page = current.page;
-                int level = current.level;
-                String contentId = page.getId();
-
-                if (!processedIds.add(contentId)) {
-                    logger.debug("Skipping already processed Confluence page {}", contentId);
-                    continue;
-                }
-
-                try {
-                    String title = page.getTitle();
-                    String safeName = deriveSafeName(title, null, index++);
-                    Path pageFolder = confluenceFolder.resolve(safeName);
-                    Files.createDirectories(pageFolder);
-
-                    // Write page text
-                    String bodyText = page.getStorage() != null && page.getStorage().getValue() != null
-                            ? page.getStorage().getValue().trim() : "";
-                    if (!bodyText.isBlank()) {
-                        String markdownText = convertHtmlToMarkdown(bodyText);
-                        Path mdFile = pageFolder.resolve(safeName + ".md");
-                        Files.write(mdFile, markdownText.getBytes(StandardCharsets.UTF_8));
-                        logger.info("Wrote Confluence page → {} ({} chars, markdown)", mdFile, markdownText.length());
-                        written++;
-
-                        // Discover related pages for the next depth level
-                        if (level < confluenceDepth) {
-                            List<Content> related = discoverRelatedPages(page, confluence, processedIds, processedUrls);
-                            for (Content relatedPage : related) {
-                                queue.add(new PageLevel(relatedPage, level + 1));
-                            }
-                        }
-                    } else {
-                        logger.warn("Confluence page '{}' has empty body, skipping text write", title);
-                    }
-
-                    // Download attachments
-                    if (confluenceAttachments && contentId != null && !contentId.isBlank()) {
-                        downloadConfluenceAttachments(confluence, contentId, pageFolder.toFile());
-                    }
-                } catch (Exception e) {
-                    logger.warn("Could not process Confluence page {} (skipping): {}", contentId, e.getMessage());
-                }
-            }
-            logger.info("Wrote {} Confluence page(s) to input/confluence/", written);
+            new ConfluencePageDownloader(confluence).downloadPages(
+                    new ArrayList<>(seedUrls), confluenceFolder.toFile(), confluenceDepth, confluenceAttachments);
         } catch (Exception e) {
             logger.warn("writeConfluencePagesFile failed (non-fatal): {}", e.getMessage());
         }
-    }
-
-    /**
-     * Discovers pages related to {@code page}: external/full Confluence URLs, children-macro children,
-     * and internal {@code ac:link} references to other pages.
-     */
-    private List<Content> discoverRelatedPages(Content page, Confluence confluence,
-                                                Set<String> alreadyQueuedIds, Set<String> alreadyProcessedUrls) {
-        List<Content> related = new ArrayList<>();
-        String storageHtml = page.getStorage() != null ? page.getStorage().getValue() : null;
-        if (storageHtml == null || storageHtml.isBlank()) {
-            return related;
-        }
-
-        // 1. External/full Confluence URLs inside the page body
-        try {
-            Set<String> urls = confluence.parseUris(storageHtml);
-            if (urls != null) {
-                for (String url : urls) {
-                    if (!alreadyProcessedUrls.add(url)) {
-                        continue;
-                    }
-                    try {
-                        Content linked = confluence.contentByUrl(url);
-                        if (linked != null && linked.getId() != null && !alreadyQueuedIds.contains(linked.getId())) {
-                            related.add(linked);
-                        }
-                    } catch (Exception e) {
-                        logger.debug("Could not resolve linked Confluence URL {}: {}", url, e.getMessage());
-                    }
-                }
-            }
-        } catch (Exception e) {
-            logger.debug("Failed to parse URIs from page {}: {}", page.getId(), e.getMessage());
-        }
-
-        // 2. Children macros
-        if (hasChildrenMacro(storageHtml)) {
-            try {
-                List<Content> children = confluence.getChildrenOfContentById(page.getId());
-                if (children != null) {
-                    for (Content child : children) {
-                        if (child != null && child.getId() != null && !alreadyQueuedIds.contains(child.getId())) {
-                            related.add(child);
-                        }
-                    }
-                }
-            } catch (Exception e) {
-                logger.warn("Could not fetch children for page {}: {}", page.getId(), e.getMessage());
-            }
-        }
-
-        // 3. Internal ac:link page references
-        List<PageLink> internalLinks = extractInternalPageLinks(storageHtml);
-        for (PageLink link : internalLinks) {
-            try {
-                Content linked;
-                if (link.spaceKey != null && !link.spaceKey.isBlank()) {
-                    linked = confluence.findContent(link.title, link.spaceKey);
-                } else {
-                    linked = confluence.findContent(link.title);
-                }
-                if (linked != null && linked.getId() != null && !alreadyQueuedIds.contains(linked.getId())) {
-                    related.add(linked);
-                }
-            } catch (Exception e) {
-                logger.debug("Could not resolve internal Confluence link '{}' (space={}): {}",
-                        link.title, link.spaceKey, e.getMessage());
-            }
-        }
-
-        return related;
-    }
-
-    /**
-     * Parses Confluence storage format HTML and returns all internal page links
-     * ({@code <ac:link><ri:page ri:content-title="..." [ri:space-key="..."]/></ac:link>}).
-     */
-    private List<PageLink> extractInternalPageLinks(String storageHtml) {
-        List<PageLink> links = new ArrayList<>();
-        if (storageHtml == null || storageHtml.isBlank()) {
-            return links;
-        }
-        try {
-            Document doc = Jsoup.parseBodyFragment(storageHtml);
-            for (Element link : doc.getElementsByTag("ac:link")) {
-                Element page = link.getElementsByTag("ri:page").first();
-                if (page != null) {
-                    String title = page.attr("ri:content-title");
-                    String spaceKey = page.attr("ri:space-key");
-                    if (title != null && !title.isBlank()) {
-                        links.add(new PageLink(title, spaceKey));
-                    }
-                }
-            }
-        } catch (Exception e) {
-            logger.debug("Failed to extract internal Confluence links: {}", e.getMessage());
-        }
-        return links;
-    }
-
-    /**
-     * Returns true if the storage format contains a {@code children} macro.
-     */
-    private boolean hasChildrenMacro(String storageHtml) {
-        if (storageHtml == null || storageHtml.isBlank()) {
-            return false;
-        }
-        try {
-            Document doc = Jsoup.parseBodyFragment(storageHtml);
-            for (Element macro : doc.getElementsByTag("ac:structured-macro")) {
-                if ("children".equalsIgnoreCase(macro.attr("ac:name"))) {
-                    return true;
-                }
-            }
-        } catch (Exception e) {
-            logger.debug("Failed to check for children macro: {}", e.getMessage());
-        }
-        return false;
-    }
-
-    /**
-     * Simple carrier for a page + its depth level in the BFS.
-     */
-    private static final class PageLevel {
-        final Content page;
-        final int level;
-
-        PageLevel(Content page, int level) {
-            this.page = page;
-            this.level = level;
-        }
-    }
-
-    /**
-     * Simple carrier for an internal Confluence page link.
-     */
-    private static final class PageLink {
-        final String title;
-        final String spaceKey;
-
-        PageLink(String title, String spaceKey) {
-            this.title = title;
-            this.spaceKey = spaceKey;
-        }
-    }
-
-    /**
-     * Downloads all attachments for a Confluence content item into {@code targetDir}
-     * via {@link Confluence#downloadPageAttachments} — the shared method used by both
-     * MermaidIndex and CLI input folder preparation.
-     */
-    private void downloadConfluenceAttachments(Confluence confluence, String contentId, File targetDir) {
-        try {
-            List<java.io.File> files = confluence.downloadPageAttachments(contentId, targetDir);
-            logger.info("Downloaded {} attachment(s) for Confluence page {} to input/confluence/",
-                    files.size(), contentId);
-        } catch (Exception e) {
-            logger.warn("Could not download attachments for content {} (skipping): {}", contentId, e.getMessage());
-        }
-    }
-
-    /** Derives a safe filesystem name from a Confluence page title or URL. */
-    private static String deriveSafeName(String title, String url, int index) {
-        if (title != null && !title.isBlank()) {
-            return title.replaceAll("[^\\w.\\-]", "_");
-        }
-        String urlPath = url.replaceAll("\\?.*", "").replaceAll("#.*", "");
-        String[] segments = urlPath.split("/");
-        String raw = segments[segments.length - 1];
-        return raw.isBlank() ? "page-" + index : raw.replaceAll("[^\\w.\\-]", "_");
-    }
-
-    /**
-     * Converts Confluence Storage Format HTML to readable Markdown.
-     * Delegates to {@link ConfluenceStorageMarkdown} which pre-processes
-     * Confluence-specific tags before passing to CopyDown.
-     */
-    private static String convertHtmlToMarkdown(String html) {
-        return ConfluenceStorageMarkdown.toMarkdown(html);
     }
 
     /**
